Moved the controller API for products to the controller folder.
